pyproject.toml
[project]
name = "mistral-ocr-mcp"
version = "0.1.0"
description = "MCP server providing Mistral OCR capabilities"
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "mcp[cli]>=0.9.0",
    "mistralai>=1.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mistral-ocr-mcp = "mistral_ocr_mcp.server:main"